Move `PartialSurfaceVertex` to `partial` module
