+++ crates/fj-kernel/src/builder/mod.rs
@@ -17,5 +17,5 @@ pub use self::{
     shell::ShellBuilder,
     sketch::SketchBuilder,
     solid::SolidBuilder,
-    vertex::{GlobalVertexBuilder, PartialSurfaceVertex},
+    vertex::GlobalVertexBuilder,
 };
