+++ crates/fj-kernel/src/partial/vertex.rs
@@ -1,9 +1,6 @@
 use fj_math::Point;
 
-use crate::{
-    builder::PartialSurfaceVertex,
-    objects::{Curve, GlobalVertex, SurfaceVertex, Vertex},
-};
+use crate::objects::{Curve, GlobalVertex, Surface, SurfaceVertex, Vertex};
 
 /// A partial [`Vertex`]
 ///
@@ -83,5 +80,70 @@ impl PartialVertex {
         let global_form = *surface_form.global_form();
 
         Vertex::new(position, curve, surface_form, global_form)
+    }
+}
+
+/// A partial [`SurfaceVertex`]
+///
+/// See [`crate::partial`] for more information.
+#[derive(Default)]
+pub struct PartialSurfaceVertex {
+    /// The position of the [`SurfaceVertex`] in the [`Surface`]
+    ///
+    /// Must be provided before [`PartialSurfaceVertex::build`] is called.
+    pub position: Option<Point<2>>,
+
+    /// The surface that the [`SurfaceVertex`] is defined in
+    ///
+    /// Must be provided before [`PartialSurfaceVertex::build`] is called.
+    pub surface: Option<Surface>,
+
+    /// The global form of the [`SurfaceVertex`]
+    ///
+    /// Can be provided, if already available, or computed from the position on
+    /// the [`Surface`].
+    pub global_form: Option<GlobalVertex>,
+}
+
+impl PartialSurfaceVertex {
+    /// Provide a position for the partial surface vertex
+    pub fn with_position(mut self, position: impl Into<Point<2>>) -> Self {
+        self.position = Some(position.into());
+        self
+    }
+
+    /// Provide a surface for the partial surface vertex
+    pub fn with_surface(mut self, surface: Surface) -> Self {
+        self.surface = Some(surface);
+        self
+    }
+
+    /// Provide a global form for the partial surface vertex
+    pub fn with_global_form(mut self, global_form: GlobalVertex) -> Self {
+        self.global_form = Some(global_form);
+        self
+    }
+
+    /// Build a full [`SurfaceVertex`] from the partial surface vertex
+    ///
+    /// # Panics
+    ///
+    /// Panics, if no position has been provided.
+    ///
+    /// Panics, if no surface has been provided.
+    pub fn build(self) -> SurfaceVertex {
+        let position = self
+            .position
+            .expect("Can't build `SurfaceVertex` without position");
+        let surface = self
+            .surface
+            .expect("Can't build `SurfaceVertex` without `Surface`");
+
+        let global_form = self.global_form.unwrap_or_else(|| {
+            GlobalVertex::builder()
+                .build_from_surface_and_position(&surface, position)
+        });
+
+        SurfaceVertex::new(position, surface, global_form)
     }
 }
